backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==1.4.52
alembic==1.11.1
pydantic==1.10.12
python-multipart==0.0.6
requests==2.31.0

# Frontend demo
streamlit==1.30.0
pandas==2.1.2
jinja2==3.1.2
reportlab==4.0.0
python-dotenv==1.0.0
scikit-learn==1.3.2
openai==1.9.0
boto3==1.28.0
sentence-transformers==2.2.2
faiss-cpu==1.8.0
llama-cpp-python==0.1.59
PyJWT==2.8.0
redis==4.6.0
rq==1.14.0
psycopg2-binary==2.9.10
pytest==7.4.0
httpx==0.24.1
passlib[bcrypt]==1.7.4
